5th update

+++ src/main/java/com/example/StudentDemo/controller/StudentController.java
@@ -3,12 +3,13 @@ package com.example.StudentDemo.controller;
 import com.example.StudentDemo.model.Student;
 import com.example.StudentDemo.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@Controller
+@RestController
 public class StudentController {
 
     private final StudentRepository studentRepository;
@@ -18,28 +19,42 @@ public class StudentController {
         this.studentRepository = studentRepository;
     }
 
-    @GetMapping
+    @RequestMapping(value = "/students",method = RequestMethod.GET)
     public List<Student> findAll() {
         return studentRepository.findAll();
     }
-    @GetMapping("/{id}")
-    public Student findById(@PathVariable Long id) {
-        return studentRepository.findById(id);
+    @RequestMapping(value = "/students/{id}",method = RequestMethod.GET)
+    public ResponseEntity<Student> findById(@PathVariable(value = "id") Long id ) {
+        Student student = studentRepository.findById(id);
+        return student != null
+                ? new ResponseEntity<>(student, HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+    @PostMapping(value = "/students/add")
+    public ResponseEntity<Void> save(@RequestBody Student student) {
+        studentRepository.save(student);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
-    @PostMapping
-    public void save(@RequestBody Student student) {
-        studentRepository.save(student);
-    }
 
     @PutMapping("/{id}")
-    public void updateStudent(@PathVariable Long id, @RequestBody Student updatedStudent) {
-        studentRepository.updateStudent(id, updatedStudent);
+    public ResponseEntity<Void> updateStudent(@PathVariable Long id, @RequestBody Student updatedStudent) {
+        if (studentRepository.findById(id) != null) {
+            studentRepository.updateStudent(id, updatedStudent);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Long id) {
-        studentRepository.deleteById(id);
+    public ResponseEntity<Void> deleteStudent(@PathVariable Long id) {
+        if (studentRepository.findById(id) != null) {
+            studentRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     // Other methods...
